+++ test/test_SeriesTransformer.py
@@ -23,6 +23,7 @@ def main():
             'Log',
             'MinMax',
             'Scale',
+            'RobustScale',
             'Sqrt',
             'Diff',
         ):